+++ test/integration/src/segments/RandomTimeEchoSegment.js
@@ -23,7 +23,7 @@ export default class RandomTimeEchoSegment extends MapSegment {
         request.get('http://localhost:8000/api/random-time-echo/' + encodeURIComponent(query.value)).end((err, res) => {
           if (res.ok) {
             var payload = JSON.parse(res.text);
-            dispatch(this._createSetResultAction(queryId, payload));
+            dispatch(this.getActionCreator().set(queryId, payload));
             resolve();
           } else {
             reject(new Error('Unable to fetch user data!'));