+++ test/integration/src/pages/redux/ClientInstantSegmentDependencies/ClientInstantSegmentDependencies.js
@@ -7,7 +7,6 @@ import ReduxStore from 'soya/lib/data/redux/ReduxStore.js';
 import RandomTimeEchoString from '../../../components/contextual/RandomTimeEchoString/RandomTimeEchoString.js';
 import FibonacciSequence from '../../../components/contextual/FibonacciSequence/FibonacciSequence.js';
 import FibonacciTotal from '../../../components/contextual/FibonacciTotal/FibonacciTotal.js';
-import { DevTools, DebugPanel, LogMonitor } from 'redux-devtools/lib/react';
 
 // TODO: Figure out how to do promise polyfill.
 import style from '../../../shared/sitewide.css';
@@ -51,9 +50,6 @@ class Component extends React.Component {
       <FibonacciSequence context={this.props.context} loadAtClient={true} number={10} />
       <h3>Recursive Segment Dependencies</h3>
       <FibonacciTotal context={this.props.context} loadAtClient={true} number={10} />
-      <DebugPanel top right bottom>
-        <DevTools store={this.props.context.store._store} monitor={LogMonitor} />
-      </DebugPanel>
     </div>;
   }
 }