+++ test/integration/src/pages/redux/SegmentCookie/SegmentCookie.js
@@ -4,7 +4,6 @@ import RenderResult from 'soya/lib/page/RenderResult';
 import ReactRenderer from 'soya/lib/page/react/ReactRenderer.js';
 import register from 'soya/lib/client/Register';
 import ReduxStore from 'soya/lib/data/redux/ReduxStore.js';
-import { DevTools, DebugPanel, LogMonitor } from 'redux-devtools/lib/react';
 
 // TODO: Figure out how to do promise polyfill.
 import style from '../../../shared/sitewide.css';
@@ -31,9 +30,6 @@ class Component extends React.Component {
       <BookingBox context={this.props.context} bookingId={29000} />
       {this.state.clientBookingBox}
       {this.state.clientErrorBookingBox}
-      <DebugPanel top right bottom>
-        <DevTools store={this.props.context.store._store} monitor={LogMonitor} />
-      </DebugPanel>
     </div>
   }