+++ test/integration/src/pages/redux/TestLocalSegment/TestLocalSegment.js
@@ -4,7 +4,6 @@ import RenderResult from 'soya/lib/page/RenderResult';
 import ReactRenderer from 'soya/lib/page/react/ReactRenderer';
 import register from 'soya/lib/client/Register';
 import ReduxStore from 'soya/lib/data/redux/ReduxStore';
-import { DevTools, DebugPanel, LogMonitor } from 'redux-devtools/lib/react';
 import smokesignals from 'soya/lib/event/smokesignals';
 
 import LyingSegment from '../../../segments/LyingSegment.js';
@@ -67,9 +66,6 @@ class Component extends React.Component {
       <ModalLayer context={this.props.context}>
         <ConfirmModal emitter={this.modalEmitter} />
       </ModalLayer>
-      <DebugPanel top right bottom>
-        <DevTools store={this.props.context.store._store} monitor={LogMonitor} />
-      </DebugPanel>
     </div>
   }